Move metrics to its own package
