+++ pkg/metrics/metrics.go
@@ -1,4 +1,4 @@
-package pstree
+package metrics
 
 import (
 	"context"
