+++ pkg/metrics/metrics_posix.go
@@ -1,7 +1,7 @@
 //go:build linux || darwin || freebsd
 // +build linux darwin freebsd
 
-package pstree
+package metrics
 
 import (
 	"syscall"
