+++ pkg/metrics/metrics_test.go
@@ -1,4 +1,4 @@
-package pstree
+package metrics
 
 import (
 	"context"
