+++ pkg/metrics/metrics_windows.go
@@ -1,7 +1,7 @@
 //go:build windows
 // +build windows
 
-package pstree
+package metrics
 
 import (
 	"errors"
